api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
weasyprint==62.3
jinja2==3.1.4
python-dotenv==1.0.1
supabase==2.9.1
google-cloud-storage==2.18.0
